requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-SQLAlchemy==3.1.1
WTForms==3.1.2
Flask-Limiter==3.8.0
Flask-Talisman==1.1.0
bcrypt==4.2.0
itsdangerous==2.2.0
requests==2.32.3
urllib3==1.26.20
importlib-metadata==6.8.0
python-dotenv==1.0.1
icalendar==5.0.12
gunicorn==22.0.0
google-generativeai==0.7.2
pytest==8.2.1

